+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/OmMetadataManagerImpl.java
@@ -170,7 +170,7 @@ public class OmMetadataManagerImpl implements OMMetadataManager {
   // non-ratis OM clusters will be binary 01 (= decimal 1)  and for ratis
   // enabled OM cluster will be binary 10 (= decimal 2). This epoch is added
   // to ensure uniqueness of objectIDs.
-  private final int omEpoch;
+  private final long omEpoch;
 
   private Map<String, Table> tableMap = new HashMap<>();
